Move the dropdown to its own file
