+++ lib/store_app/explore/search_field.dart
@@ -19,7 +19,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 import 'package:software/l10n/l10n.dart';
-import 'package:software/store_app/common/drop_down_decoration.dart';
 import 'package:software/store_app/common/snap_section.dart';
 import 'package:software/store_app/explore/explore_model.dart';
 import 'package:yaru_icons/yaru_icons.dart';
@@ -92,74 +91,6 @@ class _SearchFieldState extends State<SearchField> {
                   ),
                 )
               : null,
-        ),
-      ),
-    );
-  }
-}
-
-class SectionDropdown extends StatelessWidget {
-  const SectionDropdown({
-    // ignore: unused_element
-    super.key,
-    required this.value,
-    this.onChanged,
-  });
-
-  final SnapSection value;
-  final Function(SnapSection?)? onChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    return ClipRRect(
-      borderRadius: BorderRadius.circular(6),
-      child: Material(
-        color: Colors.transparent,
-        child: PopupMenuButton<SnapSection>(
-          tooltip: context.l10n.filterSnaps,
-          splashRadius: 20,
-          onSelected: onChanged,
-          initialValue: SnapSection.all,
-          itemBuilder: (context) {
-            return [
-              for (final section in SnapSection.values)
-                PopupMenuItem(
-                  value: section,
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.end,
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      const SizedBox(
-                        width: 5,
-                      ),
-                      SizedBox(
-                        width: 20,
-                        child: Icon(
-                          snapSectionToIcon[section],
-                          color: Theme.of(context)
-                              .colorScheme
-                              .onSurface
-                              .withOpacity(0.8),
-                        ),
-                      ),
-                      const SizedBox(
-                        width: 10,
-                      ),
-                      Expanded(
-                        flex: 1,
-                        child: Text(
-                          section.localize(
-                            context.l10n,
-                          ),
-                          overflow: TextOverflow.ellipsis,
-                        ),
-                      )
-                    ],
-                  ),
-                )
-            ];
-          },
-          child: DropDownDecoration(child: Text(value.localize(context.l10n))),
         ),
       ),
     );
